New Level + upgrade wave continue system
New scene with the routes and pathfinding set already. also fixed the wave continue script since it was lacking delay that needed. example 2 enemies in wave 1 spawned but also 5 seconds later did the 3 from wave 2 which was not exactly an error but more of a bad gameplay moment. so I added restrainer to the point when enemy count of the wave reaches 0 the next starts